requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.5
executing==0.8.3
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tornado==6.1
traitlets==5.2.2.post1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
wrapt==1.14.1
